Refactor: Move slider parts
